refactor: simplify codebase using code-simplifier agent
