+++ internal/provider/instance_types_data_source.go
@@ -157,7 +157,7 @@ func (d *instanceTypesSource) Configure(_ context.Context, req datasource.Config
 	client, ok := req.ProviderData.(*cloudriftapi.HttpClient)
 	if !ok {
 		resp.Diagnostics.AddError(
-			"Unexpected DatSource Configure Type",
+			"Unexpected DataSource Configure Type",
 			fmt.Sprintf("Expected *cloudriftapi.HttpClient, got: %T. Please report this issue to the provider developers.",
 				req.ProviderData,
 			),
