+++ internal/provider/ssh_key_data_source.go
@@ -61,7 +61,7 @@ func (d *sshKeyDataSource) Configure(_ context.Context, req datasource.Configure
 	client, ok := req.ProviderData.(*cloudriftapi.HttpClient)
 	if !ok {
 		resp.Diagnostics.AddError(
-			"Unexpected DatSource Configure Type",
+			"Unexpected DataSource Configure Type",
 			fmt.Sprintf("Expected *cloudriftapi.HttpClient, got: %T. Please report this issue to the provider developers.",
 				req.ProviderData,
 			),
